Pull payload outside the rate limiter
