+++ src/core/CMakeLists.txt
@@ -135,6 +135,7 @@ set(
   model_config_utils.cc
   model_repository_manager.cc
   numa_utils.cc
+  payload.cc
   persistent_backend_manager.cc
   pinned_memory_manager.cc
   scheduler_utils.cc
@@ -170,6 +171,7 @@ set(
   model_config_utils.h
   model_repository_manager.h
   numa_utils.h
+  payload.h
   persistent_backend_manager.h
   pinned_memory_manager.h
   rate_limiter.h
